social/requirements/production.txt
asgiref==3.2.7
Django==3.0.7
djangorestframework==3.11.0
djangorestframework_simplejwt==4.6.0
pytz==2020.1
sqlparse==0.3.1
uvicorn==0.11.7
psycopg2-binary==2.8.5
gunicorn==20.0.4
drf-yasg==1.17.1
Pillow==8.1.0